+++ src/components/QueueCreateForm.js
@@ -5,17 +5,15 @@ import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
 import Grid from '@material-ui/core/Grid';
 
+import Box from '@material-ui/core/Box';
+import Card from '@material-ui/core/Card';
+import CardHeader from '@material-ui/core/CardHeader';
+import CardContent from '@material-ui/core/CardContent';
+import Divider from '@material-ui/core/Divider';
+
 const useStyles = makeStyles((theme) => ({
   root: {
     marginTop: theme.spacing(2),
-    '& .MuiTextField-root': {
-      margin: theme.spacing(1),
-      width: 200,
-    },
-  },
-  actions: {
-    display: 'flex',
-    alignItems: 'center',
   },
 }));
 
@@ -44,22 +42,29 @@ const QueueCreateForm = ({ onSubmit }) => {
       className={classes.root}
       onSubmit={handleSubmit}
     >
-      <Typography>Add a new queue:</Typography>
-      <Grid container>
-        <Grid item>
+      <Card>
+        <CardHeader
+          title="New Queue"
+          subheader="A queue is a namespace for scheduled and repeatable tasks."
+        />
+        <Divider />
+        <CardContent>
           <TextField
+            fullWidth
+            margin="normal"
             id="queue-name"
             label="Queue Name"
             variant="outlined"
             size="small"
           />
-        </Grid>
-        <Grid item className={classes.actions}>
+        </CardContent>
+        <Divider />
+        <Box display="flex" justifyContent="flex-end" p={2}>
           <Button type="submit" variant="contained" color="primary">
             Create
           </Button>
-        </Grid>
-      </Grid>
+        </Box>
+      </Card>
     </form>
   );
 };